Add files via upload
Project/
│── app.py
│── iris_model.pkl
│
├── templates/
│     └── index.html
│     └── result.html
│
└── static/
      └── style.css

This project is a simple Machine Learning web application that predicts the species of an Iris flower based on four measurements: sepal length, sepal width, petal length, and petal width. The system uses a HTML/CSS frontend for the user interface and a Python Flask backend to process the inputs and return predictions. A trained scikit-learn machine learning model is used to classify the flower species.